Move RadioScheduleContainer styling
